// src/squads/adobo/domains/entry-exit/common/constants/enum.ts
export enum ScannerScreenStatus {
    DEFAULT = "default",
    SUCCESS = "success",
    ERROR = "error",
}

export enum CameraPermissionStatus {
    GRANTED = "granted",
    DENIED = "denied",
}

export enum ScannerResultText {
    EXIT = "exit",
    ENTRY = "entry",
    SCAN_LIMIT = "scanLimit",
    PERMISSION_DENIED = "permissionFail",
    INVALID = "invalid",
    ERROR = "error",
    TITLE = "title",
    OOPS = "oops",
    WELCOME = "welcome",
    GOODBYE = "goodbye",
    NETWORK_ISSUE = "networkIssue",
}

export enum Entities {
    ENTRY_EXIT = "entry_exit",
}